criando dtos para gasto e recebimento  e removendo os ids mockados

+++ src/main/java/br/com/fintech/controller/GastoController.java
@@ -3,6 +3,8 @@ package br.com.fintech.controller;
 import br.com.fintech.exceptions.EntityNotFoundException;
 import br.com.fintech.model.Gasto;
 import br.com.fintech.service.GastoService;
+import br.com.fintech.service.security.JwtService;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -13,65 +15,70 @@ import java.util.List;
 @RequestMapping("/api/gastos")
 public class GastoController {
     private final GastoService gastoService;
-    private static final Long MOCK_USER_ID = 1L;
+    private final JwtService jwtService;
 
-    public GastoController(GastoService gastoService) {
+    public GastoController(GastoService gastoService, JwtService jwtService) {
         this.gastoService = gastoService;
+        this.jwtService = jwtService;
+    }
+
+    private Long getUsuarioId(HttpServletRequest request) {
+        String token = request.getHeader("Authorization");
+        if (token != null && token.startsWith("Bearer ")) {
+            token = token.substring(7);
+        }
+        return jwtService.getUserIdFromToken(token);
     }
 
     @GetMapping
-    public ResponseEntity<List<Gasto>> buscarTodos() {
-        Long userId = MOCK_USER_ID;
+    public ResponseEntity<List<Gasto>> buscarTodos(HttpServletRequest request) {
+        Long userId = getUsuarioId(request);
+        if (userId == null) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
 
         List<Gasto> todosOsGastos = gastoService.findAllByOwnerId(userId);
-
         return ResponseEntity.ok(todosOsGastos);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Gasto> buscarPorId(@PathVariable("id") Long id) throws EntityNotFoundException {
-        Long userId = MOCK_USER_ID;
+    public ResponseEntity<Gasto> buscarPorId(@PathVariable("id") Long id, HttpServletRequest request)
+            throws EntityNotFoundException {
+        Long userId = getUsuarioId(request);
+        if (userId == null) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
 
         Gasto gastoPorId = gastoService.getById(id, userId);
-
         return ResponseEntity.ok(gastoPorId);
     }
 
     @PostMapping
-    public ResponseEntity<Gasto> salvar(@RequestBody Gasto gasto)
-            throws IllegalArgumentException, EntityNotFoundException
-    {
-        Long userId = MOCK_USER_ID;
+    public ResponseEntity<Gasto> salvar(@RequestBody Gasto gasto, HttpServletRequest request)
+            throws IllegalArgumentException, EntityNotFoundException {
+        Long userId = getUsuarioId(request);
+        if (userId == null) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
 
         gasto.setUsuarioId(userId);
-
         Gasto novoGasto = gastoService.insert(gasto);
-
         return ResponseEntity.status(HttpStatus.CREATED).body(novoGasto);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Gasto> atualizar(@PathVariable("id") Long id, @RequestBody Gasto gasto)
-            throws EntityNotFoundException, IllegalArgumentException
-    {
-        Long userId = MOCK_USER_ID;
+    public ResponseEntity<Gasto> atualizar(@PathVariable("id") Long id, @RequestBody Gasto gasto, HttpServletRequest request)
+            throws EntityNotFoundException, IllegalArgumentException {
+        Long userId = getUsuarioId(request);
+        if (userId == null) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
 
         gasto.setId(id);
         gasto.setUsuarioId(userId);
-
         Gasto gastoAtualizado = gastoService.update(userId, gasto);
-
         return ResponseEntity.ok(gastoAtualizado);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> remover(@PathVariable("id") Long id)
-            throws EntityNotFoundException, IllegalArgumentException
-    {
-        Long userId = MOCK_USER_ID;
+    public ResponseEntity<Void> remover(@PathVariable("id") Long id, HttpServletRequest request)
+            throws EntityNotFoundException, IllegalArgumentException {
+        Long userId = getUsuarioId(request);
+        if (userId == null) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
 
         gastoService.remove(id, userId);
-
         return ResponseEntity.noContent().build();
     }
 }
